feat: Add complete TrackTile row examples
